Provide different search modes…
Thrasher now allows focussed searching in track attributes.
In particular, it is now possible to search for a given input
in the tracks' "artist", "album", "name" or "any" of these.

The statusline now contains a track count for the matched songs
and an indication of the current search mode (see desc. above).

The default key mapping has changed to `<leader>m` and is
configurable via `g:thrashermap`.

Additionally, did some minor cleanup, making sure full function names
are used instead of abbreviated versions.